+++ types/methods/show.d.ts
@@ -1,8 +1,8 @@
 export default show;
 declare function show(dependencies: any, state: any): ({ scene: requestedScene, options }: {
     scene: string;
-    options: {
+    options?: {
         ssr: boolean;
     };
-}, ...args: any[]) => Promise<any>;
+}, ...args?: any[]) => Promise<any>;
 //# sourceMappingURL=show.d.ts.map